requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
aiogram==3.22.0
gunicorn==22.0.0
